feat(marketing): add reusable MarketingForm with multi-channel fan-out
Extract form rendering from FormSection into a standalone MarketingForm
component exported from the marketing package. The form fans out
submissions to HubSpot, Customer.io, and Notion in parallel via the
existing submitFormAction. Replace the legacy HubSpotFormEmbed iframe
on the AWS Activate page with the new form, and delete the old embed.
