+++ packages/marketing/index.ts
@@ -1,1 +1,2 @@
 export * from './src/go'
+export * from './src/forms'
