+++ packages/marketing/src/forms/index.ts
@@ -1,0 +1,6 @@
+export { default as MarketingForm } from './MarketingForm'
+export type {
+  MarketingFormCrmConfig,
+  MarketingFormField,
+  MarketingFormProps,
+} from './MarketingForm'
